pyproject.toml
[project]
name = "catchmesunsets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi>=2026.2.25",
    "cloudinary>=1.44.1",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "mysql-connector-python>=9.6.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
]
